refactor(server): type authenticator injection as a protocol
Authenticator protocol for test fakes, float recheck interval, ty-clean test annotations.

+++ src/gh_babysitter/server/config.py
@@ -11,7 +11,7 @@ class Settings:
     webhook_secret: str | None = None
     github_api_url: str = "https://api.github.com"
     auth_cache_ttl: int = 300
-    recheck_interval: int = 300
+    recheck_interval: float = 300
     ping_interval: int = 30
     queue_maxsize: int = 256
 
@@ -23,7 +23,7 @@ class Settings:
             webhook_secret=os.getenv(f"{prefix}WEBHOOK_SECRET"),
             github_api_url=os.getenv(f"{prefix}GITHUB_API_URL", "https://api.github.com"),
             auth_cache_ttl=int(os.getenv(f"{prefix}AUTH_CACHE_TTL", "300")),
-            recheck_interval=int(os.getenv(f"{prefix}RECHECK_INTERVAL", "300")),
+            recheck_interval=float(os.getenv(f"{prefix}RECHECK_INTERVAL", "300")),
             ping_interval=int(os.getenv(f"{prefix}PING_INTERVAL", "30")),
             queue_maxsize=int(os.getenv(f"{prefix}QUEUE_MAXSIZE", "256")),
         )
